Add globalargs integration tests
This required moving some test code around so that I could
access s3 helper code.
